// server/src/application/use_cases/admin/mf-management/interface/fetch-all-funds-usecase.interface.ts
import { FundListDTO } from "@application/dto/mutual-funds/mutual-fund-response.dto";
import { QueryOptions } from "mongoose";

export interface IFetchAllFundsUseCases {
    execute(data: QueryOptions): Promise<{
        data: FundListDTO[];
        total: number;
        page: number;
        limit: number;
        totalPages: number;
        totalActiveFunds?: number;
        totalInactiveFunds?: number;
        recentNavUpdates?: number
    }>
}